# Generator performance