--- a/modules/generator/generator_test.go
+++ b/modules/generator/generator_test.go
@@ -225,6 +225,7 @@ func BenchmarkCollect(b *testing.B) {
 			spanMetricsDimensions:                   []string{"k8s.cluster.name", "k8s.namespace.name"},
 			spanMetricsEnableTargetInfo:             true,
 			spanMetricsTargetInfoExcludedDimensions: []string{"excluded}"},
+			// nativeHistograms:                        overrides.HistogramMethodBoth,
 		}
 	)