--- a/modules/generator/registry/registry.go
+++ b/modules/generator/registry/registry.go
@@ -128,7 +128,7 @@ func New(cfg *Config, overrides Overrides, tenant string, appendable storage.App
 		metricFailedCollections:  metricFailedCollections.WithLabelValues(tenant),
 	}
 
-	go job(instanceCtx, r.collectMetrics, r.collectionInterval)
+	go job(instanceCtx, r.CollectMetrics, r.collectionInterval)
 	go job(instanceCtx, r.removeStaleSeries, constantInterval(5*time.Minute))
 
 	return r
@@ -201,7 +201,7 @@ func (r *ManagedRegistry) onRemoveMetricSeries(count uint32) {
 	r.metricActiveSeries.Sub(float64(count))
 }
 
-func (r *ManagedRegistry) collectMetrics(ctx context.Context) {
+func (r *ManagedRegistry) CollectMetrics(ctx context.Context) {
 	if r.overrides.MetricsGeneratorDisableCollection(r.tenant) {
 		return
 	}